pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "pyramex"
dynamic = ["version"]
description = "A Python Ramanome Analysis Toolkit for ML/DL-friendly analysis"
readme = "README_GITHUB.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Xiao Long Xia 1", email = "xiaolongxia@openclaw.cn"},
]
keywords = [
    "raman",
    "spectroscopy",
    "machine-learning",
    "deep-learning",
    "data-analysis",
    "bioinformatics",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]

dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=0.24.0",
    "matplotlib>=3.3.0",
    "plotly>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.12",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.910",
    "sphinx>=4.0",
    "sphinx-rtd-theme>=0.5",
]
ml = [
    "torch>=1.9.0",
    "tensorflow>=2.6.0",
    "umap-learn>=0.5.0",
]
gpu = [
    "cupy>=9.0",
    "numba>=0.53",
]
all = [
    "pyramex[dev,ml,gpu]",
]

[project.urls]
Homepage = "https://github.com/openclaw/pyramex"
Documentation = "https://pyramex.readthedocs.io"
Repository = "https://github.com/openclaw/pyramex"
Issues = "https://github.com/openclaw/pyramex/issues"

[tool.setuptools_scm]
write_to = "pyramex/_version.py"

[tool.setuptools.packages.find]
include = ["pyramex*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=pyramex --cov-report=term-missing"

[tool.coverage.run]
source = ["pyramex"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
